backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
shapely==2.0.2
geojson==3.1.0
python-multipart==0.0.6
numpy<2.0.0
